+++ packages/core/core/index.ts
@@ -1,3 +1,4 @@
+export type { DeepPartial } from '@ai-sdk/ui-utils';
 export * from './embed';
 export * from './generate-object';
 export * from './generate-text';
@@ -5,5 +6,5 @@ export * from './prompt';
 export * from './registry';
 export * from './tool';
 export * from './types';
-export type { DeepPartial } from '@ai-sdk/ui-utils';
 export { cosineSimilarity } from './util/cosine-similarity';
+export { jsonSchema } from './util/schema';